repackaged the files into duke package
