+++ build.gradle
@@ -1,6 +1,7 @@
 plugins {
     id 'java'
     id 'application'
+    id 'com.github.johnrengelman.shadow' version '5.1.0'
 }
 
 group 'seedu.duke'
@@ -12,7 +13,14 @@ repositories {
 
 application {
     // Change this to your main class.
-    mainClassName = "main.java.components.Duke"
+    mainClassName = "duke.Duke";
+}
+
+shadowJar {
+    archiveBaseName = "duke"
+    archiveVersion = "0.1.3"
+    archiveClassifier = null
+    archiveAppendix = null
 }
 
 run {
