+++ src/Duke.txt
@@ -1,3 +1,1 @@
-T | 1 |  read
-D | 1 |  work  |  30/08/2019 10:00
-E | 0 |  project  |  10/12/2019 10:00
+T | 1 |  work
