+++ src/main/java/duke/Parser.java
@@ -1,21 +1,21 @@
-package main.java.components;
+package duke;
 
-import main.java.command.AddCommand;
-import main.java.command.ByeCommand;
-import main.java.command.Command;
-import main.java.command.DeleteCommand;
-import main.java.command.DoneCommand;
-import main.java.command.ListCommand;
-import main.java.command.FindCommand;
+import duke.command.AddCommand;
+import duke.command.ByeCommand;
+import duke.command.Command;
+import duke.command.DeleteCommand;
+import duke.command.DoneCommand;
+import duke.command.ListCommand;
+import duke.command.FindCommand;
 
 /**
- * Parser class deals with making sense of the user command.
+ * Duke.Parser class deals with making sense of the user Duke.command.
  */
 public class Parser {
 	
 	/**
-	 * A static method that parse the input command given by the user and
-	 * deals with the command respectively.
+	 * A static method that parse the input Duke.command given by the user and
+	 * deals with the Duke.command respectively.
 	 *
 	 * @param fullCommand makes sense of the input given by the user.
 	 * @return a Command which will then be executed.
