+++ src/main/java/duke/Ui.java
@@ -1,13 +1,14 @@
-package main.java.components;
+package duke;
 
-import main.java.exception.EmptyDescriptionException;
-import main.java.exception.InvalidDescriptionException;
-import main.java.exception.InvalidInputException;
+import duke.exception.EmptyDescriptionException;
+import duke.exception.InvalidDescriptionException;
+import duke.exception.InvalidInputException;
+
 import java.text.ParseException;
 import java.util.Scanner;
 
 /**
- * Ui class deals with the interactions with the user.
+ * Duke.Ui class deals with the interactions with the user.
  */
 public class Ui {
 	
@@ -22,11 +23,12 @@ public class Ui {
 	 * enter the program. It is always shown to signal the start of the program.
 	 */
 	public void showWelcome() {
-		System.out.println("Hello! I'm Duke \nWhat can I do for you?");
+		System.out.println("Hello! I'm Duke.Duke \nWhat can I do for you?");
 	}
 	
 	/**
 	 * This method reads the input by the user.
+	 *
 	 * @return string input as keyed in by the user.
 	 */
 	public String readCommand() {
@@ -46,11 +48,9 @@ public class Ui {
 		} else if (e instanceof EmptyDescriptionException) {
 			System.out.println(String.format("OOPS!!! The description of a %s cannot be empty.", e.getMessage()));
 		} else if (e instanceof InvalidDescriptionException) {
-			System.out.println(String.format("OOPS!!! Invalid input! Make sure your %s has a description and required" +
-					" data after /at for Event or /by for Deadline.\n", e.getMessage()));
+			System.out.println(String.format("OOPS!!! Invalid input! Make sure your %s has a description and required" + " data after /at for Event or /by for Deadline.\n", e.getMessage()));
 		} else if (e instanceof ParseException) {
-			System.out.println(String.format("Please write your deadline/event date in this format: dd/MM/yyyy HH:mm," +
-					" example: 02/08/2019 14:30\n", e.getMessage()));
+			System.out.println(String.format("Please write your deadline/event date in this format: dd/MM/yyyy HH:mm," + " example: 02/08/2019 14:30\n", e.getMessage()));
 		} else {
 			System.out.println(e.getMessage());
 		}
