+++ src/main/java/duke/command/ByeCommand.java
@@ -1,7 +1,7 @@
-package main.java.command;
+package duke.command;
 
-import main.java.components.Storage;
-import main.java.components.TaskList;
+import duke.TaskList;
+import duke.Storage;
 
 /**
  * ByeCommand class which extends the abstract class Command.
@@ -13,7 +13,7 @@ public class ByeCommand extends Command {
 	/**
 	 * Class constructor
 	 *
-	 * @param command command to exit the program.
+	 * @param command Duke.command to exit the program.
 	 */
 	public ByeCommand(String command) {
 		this.command = command;
@@ -23,7 +23,7 @@ public class ByeCommand extends Command {
 	 * Method that overrides the abstract method in Command class.
 	 * Prints the goodbye message to the user.
 	 *
-	 * @param tasks ArrayList of Tasks that keep tracks of the Tasks.
+	 * @param tasks   ArrayList of Tasks that keep tracks of the Tasks.
 	 * @param storage Handles the reading and writing of the txt file.
 	 */
 	public void execute(TaskList tasks, Storage storage) {
@@ -33,7 +33,7 @@ public class ByeCommand extends Command {
 	/**
 	 * ByeCommand exits the program.
 	 *
-	 * @return True as this command end the program.
+	 * @return True as this Duke.command end the program.
 	 */
 	public boolean isExit() {
 		return true;
