+++ src/main/java/duke/command/Command.java
@@ -1,33 +1,33 @@
-package main.java.command;
+package duke.command;
 
-import main.java.components.Storage;
-import main.java.components.TaskList;
-import main.java.exception.EmptyDescriptionException;
-import main.java.exception.InvalidDescriptionException;
-import main.java.exception.InvalidInputException;
+import duke.Storage;
+import duke.TaskList;
+import duke.exception.EmptyDescriptionException;
+import duke.exception.InvalidDescriptionException;
+import duke.exception.InvalidInputException;
+
 import java.io.IOException;
 import java.text.ParseException;
 
 /**
  * An abstract Command class with abstract method execute and isExit.
- * This Command class handles the command given by the user.
+ * This Command class handles the Duke.command given by the user.
  */
 public abstract class Command {
 	
 	/**
 	 * Abstract method that is the main function of the Command class.
 	 * Respective subclass of Command (classes which inherits from Command) will have this execute
-	 * function based on their command, and carry out the respective task as per the command.
+	 * function based on their Duke.command, and carry out the respective task as per the Duke.command.
 	 *
-	 * @param tasks ArrayList of Tasks that keep tracks of the Tasks.
+	 * @param tasks   ArrayList of Tasks that keep tracks of the Tasks.
 	 * @param storage Handles the reading and writing of the txt file.
-	 * @throws EmptyDescriptionException if input does not contain description for the task.
-	 * @throws InvalidInputException if input does not qualify for any of the functions in the program.
+	 * @throws EmptyDescriptionException   if input does not contain description for the task.
+	 * @throws InvalidInputException       if input does not qualify for any of the functions in the program.
 	 * @throws InvalidDescriptionException if input does not follow the required format.
-	 * @throws IOException if storage file for tracking the TaskList is not found.
+	 * @throws IOException                 if storage file for tracking the Duke.TaskList is not found.
 	 */
-	public abstract void execute(TaskList tasks, Storage storage) throws EmptyDescriptionException,
-			InvalidInputException, InvalidDescriptionException, IOException;
+	public abstract void execute(TaskList tasks, Storage storage) throws EmptyDescriptionException, InvalidInputException, InvalidDescriptionException, IOException;
 	
 	/**
 	 * Abstract method that is most applicable to specifically the ByeCommand Class.
@@ -37,7 +37,7 @@ public abstract class Command {
 	public abstract boolean isExit();
 	
 	/**
-	 * Handles the exception thrown by respective Command execution.
+	 * Handles the Duke.exception thrown by respective Command execution.
 	 * Takes care of InvalidInputException, EmptyDescriptionException, InvalidDescriptionException,
 	 * Parse Exception and other unhandled Exceptions.
 	 *
@@ -49,11 +49,9 @@ public abstract class Command {
 		} else if (e instanceof EmptyDescriptionException) {
 			System.out.println(String.format("OOPS!!! The description of a %s cannot be empty.", e.getMessage()));
 		} else if (e instanceof InvalidDescriptionException) {
-			System.out.println(String.format("OOPS!!! Invalid input! Make sure your %s has a description and required" +
-					" data after /at for Event or /by for Deadline.\n", e.getMessage()));
+			System.out.println(String.format("OOPS!!! Invalid input! Make sure your %s has a description and required" + " data after /at for Event or /by for Deadline.\n", e.getMessage()));
 		} else if (e instanceof ParseException) {
-			System.out.println(String.format("Please write your deadline/event date in this format: dd/MM/yyyy HH:mm," +
-					" example: 02/08/2019 14:30\n", e.getMessage()));
+			System.out.println(String.format("Please write your deadline/event date in this format: dd/MM/yyyy HH:mm," + " example: 02/08/2019 14:30\n", e.getMessage()));
 		} else {
 			System.out.println(e.getMessage());
 		}
