+++ src/main/java/duke/tasks/Event.java
@@ -1,10 +1,10 @@
-package main.java.tasks;
+package duke.tasks;
 
 import java.util.Date;
 import java.text.SimpleDateFormat;
 
 /**
- * Event class that can be created by the user. Handles all tasks with events.
+ * Event class that can be created by the user. Handles all Duke.tasks with events.
  */
 public class Event extends Task{
